requirements.txt
streamlit
gspread
gspread-dataframe
google-auth-oauthlib
google-api-python-client
pandas
